requirements.txt
Flask==0.12
celery==3.1.25
requests==2.12.4
redis==2.10.5
pyjwt==1.4.2

